feat: 添加障碍物检测功能和状态机管理
- 在 `kuruma_control_dashboard.py` 中新增障碍物检测参数和配置。
- 引入状态机模块 `state_machine.py`，管理车辆的巡线和避障状态。
- 实现障碍物检测器，支持基于颜色和形状的障碍物识别。
- 更新日志系统，记录初始化和性能分析信息。
- 优化实时推理模块，集成障碍物检测和状态机控制。

+++ kuruma/core/preprocessing.py
@@ -15,7 +15,7 @@ import cv2
 # --- 🚀🚀🚀 完美匹配的预处理 (640×360 = 640×360，与Atlas完全一致) 🚀🚀🚀 ---
 # ---------------------------------------------------------------------------------
 
-def preprocess_matched_resolution(img_bgr, target_width=640, target_height=360, dtype=np.float32):
+def preprocess_matched_resolution(img_bgr, target_width=640, target_height=360, dtype=np.float16):
     """
     图片预处理，与atlas_single_image_inference.py和lane_dashboard_e2e.py完全一致
